requirements.txt
black==25.1.0
python-dotenv==1.0.1
rauth==0.7.3
